pyproject.toml
[project]
name = "fundaura-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Dhayananthan",email = "dhaya427@gmail.com"}
]

readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = ["langgraph (>=0.3.29,<0.4.0)", "pydantic (>=2.11.3,<3.0.0)", "langsmith (>=0.3.30,<0.4.0)", "langchain-anthropic (>=0.3.10,<0.4.0)", "langchain-groq (>=0.3.2,<0.4.0)", "tavily-python (>=0.5.4,<0.6.0)", "langchain-community (>=0.3.21,<0.4.0)", "mcp[cli] (>=1.6.0,<2.0.0)", "pymongo (>=4.12.0,<5.0.0)", "langchain-mcp-adapters (==0.0.9)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
